Version of the code audited by ZkSecurity
This branch contains the circuits and proof system that were audited by
ZkSecurity, only including the fixes.

On the side of the proof system it is the halo2 repo at
[this](https://github.com/midnightntwrk/halo2/commit/b9aba241e342a25fb551623b5020d63cda2d0602)
commit.

For the circuits, it is the midnight-circtuis repo at
[this](https://github.com/midnightntwrk/midnight-circuits/commit/59f630e066848e76963d0ead6aaa7b73cef417b3)
commit, with a cherry-pick of the fixes, namely fa72387, dc193735, 6326a1d1 and
fixing the dependency of the proof system to the commit above. A branch
has been created in midnight-circuits for future reference. You can
find it [here](https://github.com/midnightntwrk/midnight-circuits/tree/audited_code)

Co-authored-by: Miguel Ambrona <mac.ambrona@gmail.com>
Co-authored-by: David Nevado <davidnevadoc@users.noreply.github.com>
Co-authored-by: Itsaka Rakotonirina <itsaka.rakotonirina@gmail.com>
Co-authored-by: Zhiyong Gong <63758161+John-Gong-Math@users.noreply.github.com>
Co-authored-by: Reinhard Lüftenegger <34608483+remalue@users.noreply.github.com>